Former-commit-id: 1de6914e158be87305488fe670d0b0dbe383c7e6

+++ backend/src/main/java/com/sss/tally/api/memberpayment/dto/MemberPaymentDto.java
@@ -18,7 +18,7 @@ public class MemberPaymentDto {
 	@JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 	public static class MemberPaymentCreateDto{
 		private String memberUuid;
-		private int amount;
+		private Long amount;
 	}
 
 	@Getter
@@ -28,7 +28,7 @@ public class MemberPaymentDto {
 	@JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 	public static class MemberPaymentRespDto{
 		private String memberUuid;
-		private int amount;
+		private Long amount;
 		private String nickname;
 		private String profileImage;
 		private boolean payer;
